third_party/WebKit/LayoutTests/fast/dom/shadow/content-combinator-dynamic-attribute-change-expected.txt

PASS window.getComputedStyle(document.querySelector("#host > #child")).color is not "rgb(0, 128, 0)"
PASS window.getComputedStyle(document.querySelector("#host > #child")).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
